refactor: インフラストラクチャ層のmodelとtransformerを移動.
