+++ internal/app/api/infrastructure/api/transformer/account.go
@@ -2,7 +2,7 @@ package transformer
 
 import (
 	"github.com/atsumarukun/holos-storage-api/internal/app/api/domain/entity"
-	"github.com/atsumarukun/holos-storage-api/internal/app/api/infrastructure/model"
+	"github.com/atsumarukun/holos-storage-api/internal/app/api/infrastructure/api/model"
 )
 
 func ToAccountEntity(account *model.AccountModel) *entity.Account {
